+++ src/main/java/com/raphaelbauer/lajolla/cli/PROBind.java
@@ -1,7 +1,8 @@
+package com.raphaelbauer.lajolla.cli;
 
 import java.io.File;
 
-import org.apache.commons.cli.BasicParser;
+import org.apache.commons.cli.DefaultParser;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Options;
@@ -14,9 +15,9 @@ import com.raphaelbauer.lajolla.scoringfunctions.IScoringFunction;
 import com.raphaelbauer.lajolla.scoringfunctions.ScoreAccordingToScoringAtomDistanceOnlyIfNGramsAreSimilarFastNotIdealAndBasedOnTMSCORE;
 import com.raphaelbauer.lajolla.transformation.IFileToStringTranslator;
 import com.raphaelbauer.lajolla.transformation.IResidueToStringTransformer;
-import com.raphaelbauer.lajolla.transformation.rna.suite.DummySuiteTransformer;
-import com.raphaelbauer.lajolla.transformation.rna.suite.PDBRNATranslator;
-import com.raphaelbauer.lajolla.transformation.rna.suite.RNASuiteMatchRunner;
+import com.raphaelbauer.lajolla.transformation.protein.BetterOptimizedPhiPsiTranslator;
+import com.raphaelbauer.lajolla.transformation.protein.PDBProteinTranslator;
+import com.raphaelbauer.lajolla.transformation.protein.ProteinMatchRunner;
 import com.raphaelbauer.lajolla.utilities.SystemOutUtils;
 
 /*
@@ -24,7 +25,7 @@ import com.raphaelbauer.lajolla.utilities.SystemOutUtils;
  * @author raphael.andre.bauer@gmail.com
  *
  */
-public class RNASuite {
+public class PROBind {
 
   /**
    * Main
@@ -36,24 +37,27 @@ public class RNASuite {
 		////////////////////////////////////////////////////////////////////////
     // input files or directories:
     ////////////////////////////////////////////////////////////////////////
-    String querySuiteFile = "";
+    String queryDirOrFile = "";
 
-    String targetSuiteFile = "";
+    String targetDirOfFile = "";
 
     String outputDir = "";
 
 		////////////////////////////////////////////////////////////////////////
     // mini tuning...
     ////////////////////////////////////////////////////////////////////////
-    double minimumRefinementScore = 0.3d;
+    boolean dealWithAllModels = false;
+
+    double minimumRefinementScore = 0.2d;
 
     int numberOfResultsToWriteOut = 1;
 
 		////////////////////////////////////////////////////////////////////////
     // advanced stuff:
     ////////////////////////////////////////////////////////////////////////
-    int advancedNGramSize = 7;
+    int advancedNGramSize = 10;
 
+		//int advancedAngleDiscretion = 90;
 		//double advancedScoringRadius = 2.0d;
     EScoringFunctionRelativeSettings scoringFunctionRelativeSettings
             = EScoringFunctionRelativeSettings.basedOnSizeOfSmaller;
@@ -62,7 +66,7 @@ public class RNASuite {
 
     SystemOutUtils.showSplashScreen();
 
-    System.out.println("LaJolla RNASuite - RNA similarity screening and alignment based on RNA suite codes");
+    System.out.println("LaJolla PROBind - Protein substructure search based secondary structure");
 
     try {
 
@@ -72,18 +76,35 @@ public class RNASuite {
               .addOption("h", "help", false,
                       "Print help for this application");
 
-      opt.addOption("t", "target", true, "target suite file");
+      opt.addOption("t", "target", true, "target pdb file (directory or file)");
 
-      opt.addOption("q", "query", true, "query suite file");
+      opt.addOption("q", "query", true, "query pdb file (directory or file)");
 
       opt.addOption("o", "outputdir", true,
               "Directory where to store result files");
+
+      opt.addOption("am", "allmodels", false,
+              "Deal with all models (slower) (DEFAULT: "
+              + dealWithAllModels + ")");
+
+      opt
+              .addOption("sm", false,
+                      "Score based on smaller structure (symmetric score)");
 
       opt.addOption("ref", "minrefinementscore", true,
               "minimum refinement score needed (DEFAULT: "
               + minimumRefinementScore
               + ")");
 
+      opt.addOption("nr", "numres", true,
+              "Number of best results to write out: "
+              + numberOfResultsToWriteOut
+              + ")");
+
+//			opt.addOption("za", "anglediscretion", true,
+//					"ADVANCED: discretion angle for phi psi (DEFAULT: " 
+//					+ advancedAngleDiscretion
+//							+ ")");		
       opt.addOption("zn", "ngramsize", true,
               "ADVANCED: size of ngram window (DEFAULT: " + advancedNGramSize + ")");
 
@@ -91,51 +112,60 @@ public class RNASuite {
 //					"ADVANCED: radius taken for final scoring (DEFAULT: " +
 //					advancedScoringRadius
 //					+ ")");
-      BasicParser parser = new BasicParser();
+      DefaultParser parser = new DefaultParser();
       CommandLine cl = parser.parse(opt, args);
 
       // no arguments given => print help:
       if (args.length == 0) {
         //System.out.println("cl get arg list length : " + cl.getArgList().size());
         HelpFormatter f = new HelpFormatter();
-        f.printHelp("java -cp lajolla.jar RNASuite [options]", opt);
+        f.printHelp("java -cp lajolla.jar com.raphaelbauer.lajolla.cli.PROBind [options]", opt);
         System.exit(1);
       }
 
       if (cl.hasOption('h')) {
         HelpFormatter f = new HelpFormatter();
-        f.printHelp("java -cp lajolla.jar RNASuite [options]", opt);
+        f.printHelp("java -cp lajolla.jar com.raphaelbauer.lajolla.cli.PROBind [options]", opt);
         System.exit(1);
       }
 
+      if (cl.hasOption("sm")) {
+
+        scoringFunctionRelativeSettings
+                = EScoringFunctionRelativeSettings.basedOnSizeOfSmaller;
+
+      }
+
       if (cl.getOptionValue("t") == null) {
-        System.out.println("[ERROR] -t (target suite file) not set");
+        System.out.println("[ERROR] -t (target directory with pdb files) not set");
 
         System.exit(1);
 
       } else {
-        targetSuiteFile = cl.getOptionValue("t");
+        targetDirOfFile = cl.getOptionValue("t");
 
       }
 
       if (cl.getOptionValue("q") == null) {
-        System.out.println("[INFO] -q not set (query suite file)");
+        System.out.println("[INFO] -q not set (query directory or file)");
         System.out.println("       => using same parameters for -q and -t (all against all matching)");
 
-        querySuiteFile = targetSuiteFile;
+        queryDirOrFile = targetDirOfFile;
 
+					//System.exit(1);
       } else {
-        querySuiteFile = cl.getOptionValue("q");
+        queryDirOrFile = cl.getOptionValue("q");
 
       }
 
       if (cl.getOptionValue("o") == null) {
 
-        System.out.println("[INFO] -o not set (output directory )");
+        System.out.println("[INFO] -o not set (output directory)");
         System.out.println("       guessing a name and using that as output dir: ");
 
-        outputDir = new File(targetSuiteFile).getName() + "-" + System.currentTimeMillis() + File.separator;
+        outputDir = new File(targetDirOfFile).getName() + "-" + System.currentTimeMillis() + File.separator;
         System.out.println("       " + outputDir);
+        //System.exit(1);
 
       } else {
         outputDir = cl.getOptionValue("o");
@@ -148,6 +178,11 @@ public class RNASuite {
                 = Double.parseDouble(cl.getOptionValue("ref"));
       }
 
+      if (cl.hasOption("am")) {
+
+        dealWithAllModels = false;
+      }
+
       if (cl.hasOption("nr")) {
 
         numberOfResultsToWriteOut = Integer.parseInt(cl.getOptionValue("nr"));
@@ -156,6 +191,10 @@ public class RNASuite {
 				////////////////////////////////////////////////////////////////
       // advanced parameters:
       ////////////////////////////////////////////////////////////////
+//				if (cl.hasOption("za")) {
+//
+//					advancedAngleDiscretion = Integer.parseInt(cl.getOptionValue("za"));
+//				}
       if (cl.hasOption("zn")) {
 
         advancedNGramSize = Integer.parseInt(cl.getOptionValue("zn"));
@@ -165,36 +204,40 @@ public class RNASuite {
 //
 //					advancedScoringRadius = Integer.parseInt(cl.getOptionValue("zc"));
 //				}
-      IScoringFunction scoringFunction
-              = new ScoreAccordingToScoringAtomDistanceOnlyIfNGramsAreSimilarFastNotIdealAndBasedOnTMSCORE(
-                      scoringFunctionRelativeSettings);
-
-      IResidueToStringTransformer residueToStringTransformer
-              = new DummySuiteTransformer();
-
-      INGramTo3DTranslator nGramTo3DTranslator
-              = new NGramToStringTranslatorBasedOnSingleMatchingNGramsManyResults();
-
 				// /////////////////////////////////////////////////////////////
       // start it:
       // execute();
       // STEP 1: build the index - the sequence db:
+      IResidueToStringTransformer iResidueToStringTransformer
+              = new BetterOptimizedPhiPsiTranslator();
+
+      IScoringFunction scoringFunction
+              = new ScoreAccordingToScoringAtomDistanceOnlyIfNGramsAreSimilarFastNotIdealAndBasedOnTMSCORE(
+                      scoringFunctionRelativeSettings);
+
+      INGramTo3DTranslator nGramTo3DTranslator
+              = new NGramToStringTranslatorBasedOnSingleMatchingNGramsManyResults();
+
       IFileToStringTranslator iFileToStringTranslator
-              = new PDBRNATranslator(
+              = new PDBProteinTranslator(
+                      iResidueToStringTransformer,
+                      !dealWithAllModels,
                       scoringFunction,
-                      residueToStringTransformer,
                       nGramTo3DTranslator);
 
-      RNASuiteMatchRunner.executeSearch(
+      ProteinMatchRunner.executeSearch(
               advancedNGramSize,
               iFileToStringTranslator,
-              targetSuiteFile,
-              querySuiteFile,
+              iResidueToStringTransformer,
+              targetDirOfFile,
+              queryDirOrFile,
               outputDir + File.separator,
+              !dealWithAllModels,
               minimumRefinementScore,
               numberOfResultsToWriteOut);
 
-      long completeTimeInSecs = (System.currentTimeMillis() - beginTime) / 1000;
+      long completeTimeInSecs
+              = (System.currentTimeMillis() - beginTime) / 1000;
 
       String formattedCompleteTimeTakenInHHMMSS = String
               .format("%1$02d:%2$02d:%3$02d", completeTimeInSecs