+++ src/test/java/com/raphaelbauer/lajolla/TestingConstant.java
@@ -1,9 +1,0 @@
-
-package com.raphaelbauer.lajolla;
-
-
-public interface TestingConstant {
-  
-  String baseDirOfTestSets = "/User/ra/benchmarking/";
-  
-}