+++ src/test/java/com/raphaelbauer/lajolla/nouveau/rna/suite/RNASuiteAlignment1B23Test.java
@@ -13,9 +13,11 @@
  */
 package com.raphaelbauer.lajolla.nouveau.rna.suite;
 
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
 import java.io.File;
 
-import junit.framework.TestCase;
 import com.raphaelbauer.lajolla.ngramto3dtranslators.INGramTo3DTranslator;
 import com.raphaelbauer.lajolla.ngramto3dtranslators.NGramToStringTranslatorBasedOnSingleMatchingNGramsManyResults;
 import com.raphaelbauer.lajolla.scoringfunctions.IScoringFunction;
@@ -27,7 +29,7 @@ import com.raphaelbauer.lajolla.transformation.rna.suite.PDBRNATranslator;
 import com.raphaelbauer.lajolla.transformation.rna.suite.RNASuiteMatchRunner;
 import com.raphaelbauer.lajolla.utilities.DeleteDirRecursively;
 
-public class RNASuiteAlignment1B23Test extends TestCase {
+public class RNASuiteAlignment1B23Test {
 	
 	
 	static IScoringFunction scoringFunction = new ScoreAccordingToScoringAtomDistanceOnlyIfNGramsAreSimilarFastNotIdealAndBasedOnTMSCORE();
@@ -36,6 +38,7 @@ public class RNASuiteAlignment1B23Test extends TestCase {
 	= new NGramToStringTranslatorBasedOnSingleMatchingNGramsManyResults();
 
 	
+	@Test
 	public void testIdenticalMatchAndSearchAndAlignment() {